requirements.txt
pydub==0.25.1
ibm-watson==5.3.1
ibm-cloud-sdk-core==3.16.2
python-dotenv==1.0.0
urllib3==1.26.16
dash==2.17.1
plotly==5.23.0
waitress==3.0.0